rename serum remote to poseidon
