+++ Anchor.toml
@@ -3,7 +3,7 @@ anchor_version = "0.26.0"
 seeds = true
 
 [programs.localnet]
-serum_remote = "oBRem4fksRF79j3wRkqMHdJfTzxbEEd73JgN3mFQjSK"
+poseidon = "oBRem4fksRF79j3wRkqMHdJfTzxbEEd73JgN3mFQjSK"
 
 [registry]
 url = "https://anchor.projectserum.com"
@@ -56,4 +56,4 @@ wallet = "~/.config/solana/id.json"
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
 
 [programs.devnet]
-serum_remote = "oBRem4fksRF79j3wRkqMHdJfTzxbEEd73JgN3mFQjSK"
+poseidon = "oBRem4fksRF79j3wRkqMHdJfTzxbEEd73JgN3mFQjSK"
