+++ packages/serum-remote/package.json
@@ -14,7 +14,7 @@
     "build": "tsc",
     "build:package": "yarn clean && yarn copy-idl && rollup -c",
     "clean": "rm -rf ./dist",
-    "copy-idl": "test -e ../../target/types/serum_remote.ts && cp ../../target/types/serum_remote.ts ./src || echo \"Run anchor build first\" >&2",
+    "copy-idl": "test -e ../../target/types/poseidon.ts && cp ../../target/types/poseidon.ts ./src || echo \"Run anchor build first\" >&2",
     "prepublishOnly": "yarn copy-idl && rollup -c",
     "test": "echo \"Error: no test specified\" && exit 1",
     "type-check": "tsc --noEmit"
