+++ packages/serum-remote/src/index.ts
@@ -1,7 +1,7 @@
 import * as pdas from "./pdas";
 export * as instructions from "./instructions";
 export { pdas };
-export { SerumRemote, IDL } from "./serum_remote";
+export { Poseidon, IDL } from "./poseidon";
 export * from "./utils";
 export * from "./types";
 export * from "./dexes";
