+++ programs/poseidon/src/instructions/init_bounded_strategy.rs
