+++ programs/poseidon/src/instructions/init_bounded_strategy_v2.rs
