+++ programs/poseidon/src/macros.rs
