+++ programs/poseidon/src/state/bounded_strategy.rs
