+++ programs/poseidon/src/state/bounded_strategy_v2.rs
